Pass param for word wrap length
