+++ src/azure/Model/CodeModelTSa.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using AutoRest.Core.Model;
@@ -10,6 +9,7 @@ using AutoRest.TypeScript.Model;
 using Newtonsoft.Json;
 using AutoRest.Core.Utilities;
 using AutoRest.TypeScript.DSL;
+using AutoRest.TypeScript.Utilities;
 
 namespace AutoRest.TypeScript.Azure.Model
 {
@@ -118,7 +118,8 @@ namespace AutoRest.TypeScript.Azure.Model
         {
             TSBuilder builder = new TSBuilder();
 
-            builder.Line(LicenseHeader.GenerateLicenseHeaderTS());
+            builder.Line(LicenseHeader.GenerateLicenseHeader());
+            builder.Line();
 
             CompositeTypeTS[] orderedMapperTemplateModels = OrderedMapperTemplateModels.ToArray();
 
@@ -160,8 +161,8 @@ namespace AutoRest.TypeScript.Azure.Model
         public override string GenerateModelIndex()
         {
             TSBuilder builder = new TSBuilder();
-
-            builder.Line(LicenseHeader.GenerateLicenseHeaderTS());
+            builder.Line(LicenseHeader.GenerateLicenseHeader());
+            builder.Line();
             builder.Line(ConstructRuntimeImportForModelIndex());
             if (ContainsDurationPropertyInModels() || IsAnyModelInheritingFromRequestOptionsBase() || MethodsWithCustomResponseType.Any())
             {
