+++ src/azure/Templates/AzureServiceClientTemplate.cshtml
@@ -1,8 +1,9 @@
 ﻿@using System.Linq
 @using AutoRest.TypeScript
 @using AutoRest.TypeScript.Azure.Model
+@using AutoRest.TypeScript.Utilities
 @inherits AutoRest.Core.Template<CodeModelTSa>
-@(LicenseHeader.GenerateLicenseHeaderTS())
+@(LicenseHeader.GenerateLicenseHeader(AutoRest.Core.Settings.DefaultMaximumCommentColumns))
 @EmptyLine
 @(Model.GenerateAzureServiceClientImports())
 @EmptyLine
