+++ src/vanilla/Templates/MethodGroupIndexTemplate.cshtml
@@ -1,7 +1,8 @@
 ﻿@using System.Linq
 @using AutoRest.Core.Utilities
+@using AutoRest.TypeScript.Utilities
 @inherits AutoRest.Core.Template<AutoRest.TypeScript.Model.CodeModelTS>
-@(LicenseHeader.GenerateLicenseHeaderTS())
+@(LicenseHeader.GenerateLicenseHeader(AutoRest.Core.Settings.DefaultMaximumCommentColumns))
 @EmptyLine
 @foreach (var methodGroup in Model.MethodGroupModels)
 {
