+++ src/vanilla/Templates/MethodGroupTemplate.cshtml
@@ -1,6 +1,7 @@
 ﻿@using AutoRest.TypeScript
+@using AutoRest.TypeScript.Utilities
 @inherits AutoRest.Core.Template<AutoRest.TypeScript.Model.MethodGroupTS>
-@(LicenseHeader.GenerateLicenseHeaderTS())
+@(LicenseHeader.GenerateLicenseHeader(AutoRest.Core.Settings.DefaultMaximumCommentColumns))
 @EmptyLine
 @(Model.GenerateMethodGroupImports())
 @EmptyLine
