+++ src/vanilla/Templates/ParameterTemplate.cshtml
@@ -1,9 +1,12 @@
 ﻿@using System.Linq
 @using System.Text.RegularExpressions
 @using AutoRest.TypeScript.vanilla.Templates
+@using AutoRest.TypeScript.Utilities
 @inherits AutoRest.Core.Template<AutoRest.TypeScript.Model.CodeModelTS>
-@(LicenseHeader.GenerateLicenseHeaderTS())
+@(LicenseHeader.GenerateLicenseHeader(AutoRest.Core.Settings.DefaultMaximumCommentColumns))
 @EmptyLine
+
 import * as msRest from "@@azure/ms-rest-js";
 @EmptyLine
+
 @Model.GenerateParameterMappers()
